Rearrange math code to all be in one file. Easier to hunt down.
