+++ StereoKitC/StereoKitC.vcxproj.filters
@@ -58,9 +58,6 @@
     <ClCompile Include="systems\line_drawer.cpp">
       <Filter>systems</Filter>
     </ClCompile>
-    <ClCompile Include="intersect.cpp">
-      <Filter>src</Filter>
-    </ClCompile>
     <ClCompile Include="color.cpp">
       <Filter>src</Filter>
     </ClCompile>
