+++ StereoKitC/sk_math.h
@@ -28,7 +28,10 @@ inline float math_ease_hop      (float a, float peak, float t) { return a+(peak-
 // swing - rotation around axis that is perpendicular to "direction" vector
 void quat_decompose_swing_twist(quat rotation, vec3 direction, quat *out_swing, quat *out_twist);
 
-vec3 bounds_corner (const bounds_t &bounds, int32_t index8);
+bool32_t bounds_ray_intersect_dist(bounds_t bounds, ray_t ray, float* out_distance);
+vec3     bounds_corner            (bounds_t bounds, int32_t index8);
+float    bounds_sdf_manhattan     (bounds_t bounds, vec3 sample_point);
+
 vec3 math_cubemap_corner(int i);
 
 } // namespace sk
